+++ src/lib/ngx-nav-drawer/ngx-nav-drawer.component.ts
@@ -8,12 +8,18 @@ import {
   ViewEncapsulation,
   ChangeDetectorRef,
   PLATFORM_ID,
-  Inject
+  Inject,
+  HostListener,
+  ViewChild,
+  ElementRef,
+  OnChanges,
+  SimpleChanges
 } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
 import { isPlatformBrowser, isPlatformServer } from '@angular/common';
-
 import { fromEvent } from 'rxjs/observable/fromEvent';
+import { Subscription } from 'rxjs/Subscription';
+
+import { trapFocus, KeyCodes } from './util';
 
 @Component({
   selector: 'ngx-nav-drawer',
@@ -21,19 +27,22 @@ import { fromEvent } from 'rxjs/observable/fromEvent';
   styleUrls: ['./ngx-nav-drawer.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class NgxNavDrawerComponent implements OnDestroy, OnInit {
+export class NgxNavDrawerComponent implements OnChanges, OnDestroy, OnInit {
   @Output() openChange = new EventEmitter<boolean>();
   @Input()
   set open(value: boolean) {
     this.show = value;
     this.openChange.emit(this.show);
   }
+
   @Input() fixed = false;
   @Input() fixedAtWidth = '1024px';
+  @ViewChild('nav') nav: ElementRef;
 
   show = false;
   fixedMode = false;
   private subscription: Subscription;
+  private lastFocusedElement: HTMLElement;
 
   constructor(
     @Inject(PLATFORM_ID) private platformId: any,
@@ -49,9 +58,25 @@ export class NgxNavDrawerComponent implements OnDestroy, OnInit {
     }
   }
 
+  @HostListener('document:click', ['$event'])
+  rootClick(event) {
+    this.lastFocusedElement = event.target;
+  }
+
+  ngOnChanges() {
+    this.focus();
+  }
+
   ngOnDestroy() {
     if (this.subscription) {
       this.subscription.unsubscribe();
+    }
+  }
+
+  @HostListener('window:keyup', ['$event'])
+  outerClick(event) {
+    if (event.keyCode === KeyCodes.Escape && this.show === true) {
+      this.toggle();
     }
   }
 
@@ -74,6 +99,16 @@ export class NgxNavDrawerComponent implements OnDestroy, OnInit {
     if (!this.fixedMode) {
       this.show = !this.show;
       this.openChange.emit(this.show);
+    }
+
+    this.focus();
+  }
+
+  focus() {
+    if (this.show) {
+      trapFocus(this.nav.nativeElement);
+    } else if (this.lastFocusedElement) {
+      this.lastFocusedElement.focus();
     }
   }
 }