+++ src/lib/ngx-code-example/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 0.0.1
+- make prismjs a peer dep
+
 ## 0.0.0
 - initial release