+++ src/lib/ngx-modal/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 0.1.0
+- improve focus events
+
 ## 0.0.0
 - initial release