+++ src/lib/ngx-nav-drawer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.3
+- improve focus events
+- fixed styles for nav buttons
+
 ## 0.0.3
 - fix nav overflow